chore: move several classes and helpers to production
