+++ SmartCityExploration/SmartCityExploration/API/Model/RemoteCityItem.swift
@@ -20,3 +20,9 @@ struct RemoteCityItem: Codable, Equatable {
         let lon, lat: Double
     }
 }
+
+extension RemoteCityItem {
+    var toLocal: CityItem {
+        .init(country: country, name: name, id: id, latitude: coord.lat, longitude: coord.lon, isFavorite: false)
+    }
+}
