+++ SmartCityExploration/SmartCityExplorationTests/Feature/ViewModel/CitiesViewModelTests.swift
@@ -9,175 +9,6 @@
 import XCTest
 import Combine
 
-protocol SearchHelperProtocol {
-    func initiate(with items: [CityItem])
-    func search(text: String) -> [CityItem]
-}
-
-final class SearchHelper: SearchHelperProtocol {
-    private let trie = Trie()
-
-    func initiate(with items: [CityItem]) {
-        items.forEach { trie.insert($0) }
-    }
-    
-    func search(text: String) -> [CityItem] {
-        return trie.searchPrefix(text)
-    }
-}
-
-class TrieNode {
-    var children: [Character: TrieNode] = [:]
-    var items: [CityItem] = []
-}
-
-class Trie {
-    private let root = TrieNode()
-
-    // Insert an item into the Trie
-    func insert(_ item: CityItem) {
-        var node = root
-        for char in item.name.lowercased() {
-            if node.children[char] == nil {
-                node.children[char] = TrieNode()
-            }
-            node = node.children[char]!
-        }
-        node.items.append(item) // Store the item at the end node
-    }
-
-    // Search for all items with a given prefix
-    func searchPrefix(_ prefix: String) -> [CityItem] {
-        var node = root
-        var result: [CityItem] = []
-        let prefix = prefix.lowercased()
-
-        // Traverse to the node representing the prefix
-        for char in prefix {
-            guard let child = node.children[char] else { return result }
-            node = child
-        }
-
-        // Perform DFS to find all items with the given prefix
-        dfs(node, &result)
-        return result
-    }
-
-    private func dfs(_ node: TrieNode, _ result: inout [CityItem]) {
-        // Add all items at the current node
-        result.append(contentsOf: node.items)
-        // Traverse all children
-        for (_, child) in node.children {
-            dfs(child, &result)
-        }
-    }
-}
-
-
-final class CitiesViewModel {
-    private let apiClient: CitiesAPIClientProtocol
-    private let persistentStore: CitiesPersistentStore
-    private let searchHelper: SearchHelperProtocol
-
-    init(apiClient: CitiesAPIClientProtocol, persistentStore: CitiesPersistentStore, searchHelper: SearchHelperProtocol) {
-        self.apiClient = apiClient
-        self.persistentStore = persistentStore
-        self.searchHelper = searchHelper
-    }
-
-    enum UIState: Equatable {
-        case loading
-        case idle(items: [CityItem])
-        case error
-    }
-
-    @Published var state: UIState = .loading
-
-    func performSearch(of text: String) {
-        let filteredItems = searchHelper.search(text: text)
-        state = .idle(items: filteredItems.sorted(by: { $0.name < $1.name }))
-    }
-    
-    /// This function is only executed once
-    func retrieveCities() {
-        Task {
-            let result = await persistentStore.retrieve()
-
-            switch result {
-            case .success(let items) where items == .none:
-                await performRemoteRetrieveAndSave()
-            case .success(let items):
-                guard let items else {
-                    state = .error
-                    return
-                }
-                state = .idle(items: items.sorted(by: { $0.name < $1.name }))
-                searchHelper.initiate(with: items)
-            case .failure:
-                state = .error
-
-            }
-        }
-    }
-
-    private func performRemoteRetrieveAndSave() async {
-        guard let remoteItems = await retrieveFromNetwork() else {
-            state = .error
-            return
-        }
-
-        do {
-            try await insertToStore(items: remoteItems.map { $0.toLocal })
-        } catch {
-            state = .error
-        }
-
-        let cities = await retrieveFromStore()
-
-        state = .idle(items: cities)
-        searchHelper.initiate(with: cities)
-    }
-
-    private func retrieveFromNetwork() async -> [RemoteCityItem]? {
-        let result = await apiClient.retrieveCities()
-        switch result {
-        case .success(let cities):
-            return cities
-        case .failure:
-            state = .error
-            return nil
-        }
-    }
-
-    private func insertToStore(items: [CityItem]) async throws {
-        let result = await persistentStore.insert(items: items, timestamp: .now)
-        switch result {
-        case .success:
-            break
-        case .failure(let error):
-            throw error
-        }
-    }
-
-    private func retrieveFromStore() async -> [CityItem] {
-        let result = await persistentStore.retrieve()
-        switch result {
-        case .success(let items):
-            guard let items else { return [] }
-            return items
-        case .failure:
-            return []
-        }
-    }
-
-}
-
-private extension RemoteCityItem {
-    var toLocal: CityItem {
-        .init(country: country, name: name, id: id, latitude: coord.lat, longitude: coord.lon, isFavorite: false)
-    }
-}
-
 final class CitiesViewModelTests: XCTestCase {
     private var cancellables: Set<AnyCancellable> = []
